+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 #### 5.0.0-beta.16
 * Make Adapter.getLogger a verbose log operation
 * Fix #347 - BufferFull result from Logary proper, through Facade cast exn
+* Fix #346 — Stackdriver not sending Logger.name — now sent as `labels.logger`.
 
 #### 5.0.0-beta.15
 * Fix #341 — template from context as well as fields